TS to favorites.
